Move the session-free GCI library tests under gciLibrary
Pure relocation ahead of the rewrite that follows, so git records these as
renames and their history survives: the three session-free binding test
files move from the excluded `gci` vitest project into a co-located
gciLibrary/__tests__ folder and pick up the .integration.test.ts suffix.
Only the gciTestConfig import path changes, to keep the tree compiling.
